refactor(dropdown element): create dropdown-list-element from dropdown element
dropdown-element now only encapsulates opening/closing of dropdown menu, while dropdown-list-element
inherits dropdown-element to provide a more specific version of dropdown-element where menu items
are displayed in a list. datetime-picker extends dropdown-element directly, while combobox,
dropdown, mainnav extends dropdown-list-element.

BREAKING CHANGE: Refactoring of dropdown element
