+++ mocks/dropdown.ts
@@ -1,15 +1,10 @@
 import { html } from "lit";
 import { customElement, queryAsync } from "lit/decorators.js";
 import { ref } from "lit/directives/ref.js";
-import { DropdownElement } from "../src/base/dropdown-element";
 import genId from "../src/utils/generateId";
+import { DropdownListElement } from "../src/base/dropdown-list-element";
 @customElement("mock-dropdown")
-export class MockDropdown extends DropdownElement {
-  // Reset menu for keyboard interaction when menu is hide. Should use this for dropdown components that required scrolling of menu using arrow down or up key
-  connectedCallback(): void {
-    super.connectedCallback();
-    this.addEventListener("sgds-hide", this._resetMenu);
-  }
+export class MockDropdown extends DropdownListElement {
   // To wait for the dropdownRef to update finish before toggling show when menuIsOpen is true
   @queryAsync("button")
   button: Promise<HTMLButtonElement>;
@@ -21,6 +16,7 @@ export class MockDropdown extends DropdownElement {
       this.bsDropdown.show();
     }
   }
+
   render() {
     return html`
       <div>
@@ -28,13 +24,13 @@ export class MockDropdown extends DropdownElement {
           ?disabled=${this.disabled}
           aria-expanded="${this.menuIsOpen}"
           ${ref(this.myDropdown)}
-          @click=${() => this._onClickButton()}
+          @click=${() => this.toggleMenu()}
           id=${genId("dropdown", "button")}
         >
           Mock Dropdown
         </button>
         <ul class="dropdown-menu" role="menu" part="menu">
-          <slot id="default" @click=${this._handleSelectSlot}></slot>
+          <slot id="default" @click=${this.handleSelectSlot}></slot>
         </ul>
       </div>
     `;
