+++ src/base/dropdown-element.ts
@@ -1,17 +1,14 @@
 import type { StrictModifiers } from "@popperjs/core";
 import * as Popper from "@popperjs/core";
 import { Dropdown } from "bootstrap";
-import { property, query, state } from "lit/decorators.js";
+import { property, state } from "lit/decorators.js";
 import { Ref, createRef } from "lit/directives/ref.js";
-import { SgdsDropdownItem } from "../components/Dropdown/sgds-dropdown-item";
-import genId from "../utils/generateId";
 import mergeDeep from "../utils/mergeDeep";
 import SgdsElement from "./sgds-element";
 
 const ARROW_DOWN = "ArrowDown";
 const ARROW_UP = "ArrowUp";
 const ESC = "Escape";
-const ENTER = "Enter";
 
 export type DropDirection = "left" | "right" | "up" | "down";
 
@@ -25,12 +22,11 @@ export type DropDirection = "left" | "right" | "up" | "down";
 
 export class DropdownElement extends SgdsElement {
   static styles = SgdsElement.styles;
-  /**@internal */
-  @query("ul.dropdown-menu") menu: HTMLUListElement;
+
   /** @internal */
-  myDropdown: Ref<HTMLElement> = createRef();
+  protected myDropdown: Ref<HTMLElement> = createRef();
   /** @internal */
-  bsDropdown: Dropdown = null;
+  protected bsDropdown: Dropdown = null;
 
   /** @internal Controls auto-flipping of menu */
   @property({ type: Boolean, state: true })
@@ -48,12 +44,12 @@ export class DropdownElement extends SgdsElement {
   @property({ type: Object })
   popperOpts = {};
 
-  /** @internal Forwards value to id attribute of toggle button of Dropdown. An unique id generated by default */
-  protected togglerId: string = genId("dropdown", "button");
-
-  /** Sets the text content of Dropdown button */
-  @property({ type: String })
-  togglerText = "";
+  /** @internal */
+  @state()
+  dropdownConfig: Partial<Popper.Options>;
+  /** @internal */
+  @state()
+  modifierOpt: StrictModifiers[] = [];
 
   /** When true, dropdown menu shows on first load */
   @property({ type: Boolean, reflect: true })
@@ -67,18 +63,19 @@ export class DropdownElement extends SgdsElement {
   @property({ type: Boolean, reflect: true })
   disabled = false;
 
-  /** @internal */
-  @state()
-  nextDropdownItemNo = 0;
-  /** @internal */
-  @state()
-  prevDropdownItemNo = -1;
-  /** @internal */
-  @state()
-  dropdownConfig: Partial<Popper.Options>;
-  /** @internal */
-  @state()
-  modifierOpt: StrictModifiers[] = [];
+  connectedCallback() {
+    super.connectedCallback();
+
+    if (this.close !== "inside") {
+      document.addEventListener("click", (event: MouseEvent) => this._handleClickOutOfElement(event, this));
+    }
+  }
+
+  disconnectedCallback() {
+    super.disconnectedCallback();
+
+    document.removeEventListener("click", (event: MouseEvent) => this._handleClickOutOfElement(event, this));
+  }
 
   firstUpdated() {
     this.bsDropdown = new Dropdown(this.myDropdown.value, {
@@ -140,123 +137,43 @@ export class DropdownElement extends SgdsElement {
       this.emit("sgds-after-hide");
     });
 
-    this.addEventListener("keydown", this._handleKeyboardEvent);
-    if (this.close !== "inside") {
-      this.addEventListener("blur", e => {
-        // when clicking outside of the dropdown component, it becomes null, hide the dropdown
-        return e.relatedTarget === null ? this.bsDropdown.hide() : e.stopPropagation();
-      });
-      addEventListener("click", e => this._handleClickOutOfElement(e, this));
-    }
+    this.addEventListener("keydown", this._handleKeyboardMenuEvent);
   }
+
   /** When invoked, opens the dropdown menu */
-  public showMenu() {
+  showMenu() {
     this.bsDropdown.show();
   }
 
   /** When invoked, hides the dropdown menu */
-  public hideMenu() {
+  hideMenu() {
     this.bsDropdown.hide();
   }
 
-  _onClickButton() {
+  toggleMenu() {
     this.bsDropdown.toggle();
   }
 
-  _resetMenu() {
-    this.nextDropdownItemNo = 0;
-    this.prevDropdownItemNo = -1;
-    // reset the tabindex
-    const items = this._getMenuItems();
-    items.forEach(i => {
-      i.removeAttribute("tabindex");
-    });
-  }
-  _getMenuItems(): SgdsDropdownItem[] {
-    // for case when default slot is used e.g. dropdown, mainnavdropdown
-    if (this.shadowRoot.querySelector("slot#default")) {
-      return (this.shadowRoot.querySelector("slot#default") as HTMLSlotElement)?.assignedElements({
-        flatten: true
-      }) as SgdsDropdownItem[];
-    }
-
-    // for case when there is no slot e.g. combobox
-    if (this.menu.hasChildNodes()) {
-      const menuItems = this.menu.children;
-
-      return [...menuItems] as SgdsDropdownItem[];
-    }
-  }
-
-  _getActiveMenuItems(): SgdsDropdownItem[] {
-    return this._getMenuItems().filter(item => !item.disabled);
-  }
-
-  _setMenuItem(currentItemIdx: number, isArrowDown = true) {
-    const items = this._getActiveMenuItems();
-    if (items.length === 0) return;
-    const item = items[currentItemIdx];
-    this.nextDropdownItemNo = currentItemIdx + 1;
-    this.prevDropdownItemNo = currentItemIdx - 1 < 0 ? items.length - 1 : currentItemIdx - 1;
-    let activeItem: SgdsDropdownItem;
-    if (item.disabled) {
-      return this._setMenuItem(isArrowDown ? this.nextDropdownItemNo : this.prevDropdownItemNo);
-    } else activeItem = item;
-
-    // focus or blur items depending on active or not
-    items.forEach(i => {
-      i.setAttribute("tabindex", i === activeItem ? "0" : "-1");
-      i === activeItem && i.focus();
-    });
-  }
-
-  _handleSelectSlot(e: KeyboardEvent | MouseEvent) {
-    const items = this._getActiveMenuItems();
-    const currentItemNo = items.indexOf(e.target as SgdsDropdownItem);
-    this.nextDropdownItemNo = currentItemNo + 1;
-    this.prevDropdownItemNo = currentItemNo <= 0 ? items.length - 1 : currentItemNo - 1;
-
-    /** Emitted event from SgdsDropdown element when a slot item is selected */
-    const selectedItem = e.target as SgdsDropdownItem;
-    if (!selectedItem.disabled) {
-      this.emit("sgds-select");
-      this.close !== "outside" && this.bsDropdown.hide();
-    } else return;
-  }
-  _handleKeyboardEvent(e: KeyboardEvent) {
-    const menuItems = this._getActiveMenuItems();
+  private _handleKeyboardMenuEvent(e: KeyboardEvent) {
     switch (e.key) {
       case ARROW_DOWN:
         e.preventDefault();
-        if (!this.menuIsOpen) return this.bsDropdown.show();
-        if (this.nextDropdownItemNo === menuItems.length) {
-          return this._setMenuItem(0);
-        } else {
-          return this._setMenuItem(this.nextDropdownItemNo > 0 ? this.nextDropdownItemNo : 0);
-        }
+        if (!this.menuIsOpen) return this.showMenu();
+        break;
       case ARROW_UP:
         e.preventDefault();
-        if (!this.menuIsOpen) return this.bsDropdown.show();
-        if (this.prevDropdownItemNo < 0) {
-          return this._setMenuItem(menuItems.length - 1, false);
-        } else {
-          return this._setMenuItem(this.prevDropdownItemNo, false);
-        }
-      case ESC:
-        return this.bsDropdown.hide();
-      case ENTER:
-        if (menuItems.includes(e.target as SgdsDropdownItem)) {
-          return this._handleSelectSlot(e);
-        }
+        if (!this.menuIsOpen) return this.showMenu();
         break;
+      case ESC:
+        return this.hideMenu();
       default:
         break;
     }
   }
 
-  _handleClickOutOfElement(e: MouseEvent, self: DropdownElement) {
+  private _handleClickOutOfElement(e: MouseEvent, self: DropdownElement) {
     if (!e.composedPath().includes(self)) {
-      this.bsDropdown.hide();
+      this.hideMenu();
     }
   }
 }
