+++ src/components/Dropdown/sgds-dropdown.ts
@@ -4,6 +4,8 @@ import { property, queryAsync } from "lit/decorators.js";
 import { ref } from "lit/directives/ref.js";
 import { DropdownElement } from "../../base/dropdown-element";
 import { SgdsButton } from "../Button";
+import { DropdownListElement } from "../../base/dropdown-list-element";
+import genId from "../../utils/generateId";
 
 export type DropDirection = "left" | "right" | "up" | "down";
 export type DropdownButtonVariant =
@@ -22,7 +24,7 @@ export type DropdownButtonVariant =
  *
  * @csspart menu - The dropdown's menu (ul element)
  */
-export class SgdsDropdown extends ScopedElementsMixin(DropdownElement) {
+export class SgdsDropdown extends ScopedElementsMixin(DropdownListElement) {
   static styles = [DropdownElement.styles];
   static get scopedElements() {
     return {
@@ -42,10 +44,13 @@ export class SgdsDropdown extends ScopedElementsMixin(DropdownElement) {
     ];
   }
 
-  connectedCallback(): void {
-    super.connectedCallback();
-    this.addEventListener("sgds-hide", this._resetMenu);
-  }
+  /** @internal Forwards value to id attribute of toggle button of Dropdown. An unique id generated by default */
+  private togglerId: string = genId("dropdown", "button");
+
+  /** Sets the text content of Dropdown button */
+  @property({ type: String })
+  private togglerText = "";
+
   /** Controls auto-flipping of menu */
   @property({ type: Boolean, reflect: true, state: false })
   noFlip = false;
@@ -70,7 +75,7 @@ export class SgdsDropdown extends ScopedElementsMixin(DropdownElement) {
     super.firstUpdated();
     if (this.menuIsOpen) {
       await this.dropdownRef;
-      this.bsDropdown.show();
+      this.showMenu();
     }
   }
   render() {
@@ -81,7 +86,7 @@ export class SgdsDropdown extends ScopedElementsMixin(DropdownElement) {
           ?disabled=${this.disabled}
           aria-expanded="${this.menuIsOpen}"
           ${ref(this.myDropdown)}
-          @click=${() => this._onClickButton()}
+          @click=${() => this.toggleMenu()}
           id=${this.togglerId}
         >
           ${this.togglerText}
@@ -100,7 +105,7 @@ export class SgdsDropdown extends ScopedElementsMixin(DropdownElement) {
           </svg>
         </sgds-button>
         <ul class="dropdown-menu" role="menu" part="menu">
-          <slot id="default" @click=${this._handleSelectSlot}></slot>
+          <slot id="default" @click=${this.handleSelectSlot}></slot>
         </ul>
       </div>
     `;
