+++ src/components/Mainnav/sgds-mainnav-dropdown.ts
@@ -1,28 +1,34 @@
 import { html } from "lit";
-import { queryAsync } from "lit/decorators.js";
+import { property, queryAsync } from "lit/decorators.js";
 import { ref } from "lit/directives/ref.js";
 import { DropdownElement } from "../../base/dropdown-element";
 import styles from "./sgds-mainnav-dropdown.scss";
+import { DropdownListElement } from "../../base/dropdown-list-element";
+import genId from "../../utils/generateId";
+
 /**
  * @slot default - The menu items. Pass in sgds-dropdown-item as the menu items
  * @slot toggler - The content of the toggler to pass in html content. `togglerText` property takes precedence over this slot when both are defined.
  */
-export class SgdsMainnavDropdown extends DropdownElement {
+export class SgdsMainnavDropdown extends DropdownListElement {
   static styles = [DropdownElement.styles, styles];
 
-  connectedCallback(): void {
-    super.connectedCallback();
-    this.addEventListener("sgds-hide", this._resetMenu);
-  }
   /**@internal */
   @queryAsync("a")
-  dropdownRef: Promise<HTMLAnchorElement>;
+  private dropdownRef: Promise<HTMLAnchorElement>;
+
+  /** @internal Forwards value to id attribute of toggle button of Dropdown. An unique id generated by default */
+  private togglerId: string = genId("dropdown", "button");
+
+  /** Sets the text content of Dropdown button */
+  @property({ type: String })
+  private togglerText = "";
 
   async firstUpdated() {
     super.firstUpdated();
     if (this.menuIsOpen) {
       await this.dropdownRef;
-      this.bsDropdown.show();
+      this.showMenu();
     }
   }
 
@@ -34,7 +40,7 @@ export class SgdsMainnavDropdown extends DropdownElement {
           ?disabled=${this.disabled}
           aria-expanded="${this.menuIsOpen}"
           ${ref(this.myDropdown)}
-          @click=${() => this._onClickButton()}
+          @click=${() => this.toggleMenu()}
           id=${this.togglerId}
           tabindex=${this.disabled ? "-1" : "0"}
           role="button"
@@ -55,7 +61,7 @@ export class SgdsMainnavDropdown extends DropdownElement {
           </svg>
         </a>
         <ul class="dropdown-menu" role="menu" part="menu">
-          <slot id="default" @click=${this._handleSelectSlot}></slot>
+          <slot id="default" @click=${this.handleSelectSlot}></slot>
         </ul>
       </li>
     `;
